% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{agcd}
\alias{agcd}
\title{Greatest Common Divisor of Rational Numbers}
\usage{
agcd(x, precision)
}
\arguments{
\item{x}{Vector of rational numbers}

\item{precision}{Precision value for creating rational fractions}
}
\value{
The greatest common divisor of the rational numbers
}
\description{
See equation 15 in "Non-Integer Arrays for Array Signal Processing"
by Kulkarni and Vaidyanathan (2022)
}
\details{
See Huang et al. (2014)

See Amman and Das (2001)
}
