% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{alcm}
\alias{alcm}
\title{alcm}
\usage{
alcm(x, deviation)
}
\arguments{
\item{x}{Vector of integers to approximate}

\item{deviation}{Deviation for estimating least common multiples}
}
\value{
Approximate least common multiple
}
\description{
Approximate least common multiple
}
