% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{analyze_harmonics}
\alias{analyze_harmonics}
\title{analyze_harmonics}
\usage{
analyze_harmonics(x)
}
\arguments{
\item{x}{Chord frequencies}
}
\value{
A data frame of frequencies and pseudo_octaves
}
\description{
Determine pseudo octave of all frequencies relative to lowest frequency
}
