% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{approximate_harmonics}
\alias{approximate_harmonics}
\title{approximate_harmonics}
\usage{
approximate_harmonics(x, approximate_lcm_sd)
}
\arguments{
\item{x}{Chord frequencies}

\item{approximate_lcm_sd}{Deviation for estimating least common multiples}
}
\value{
A double of the best guess of the pseudo octave
}
\description{
Determine pseudo octave of all frequencies relative to lowest frequency
}
