% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{approximate_lcm_deviation}
\alias{approximate_lcm_deviation}
\title{Default Approximate Least Common Multiple Deviation}
\usage{
approximate_lcm_deviation()
}
\description{
Default deviation for approximating the Least Common Multiple (LCM)
}
\details{
'
}
