% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{approximate_pseudo_octave}
\alias{approximate_pseudo_octave}
\title{approximate_pseudo_octave}
\usage{
approximate_pseudo_octave(approximate_harmonics)
}
\arguments{
\item{approximate_harmonics}{List of candidate pseudo octaves}
}
\value{
A data frame of rational numbers and metadata
}
\description{
Finds the pseudo octave from approximate harmonics.
}
