% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{compute_fundamental_cycle}
\alias{compute_fundamental_cycle}
\title{Compute the cycle length of a complex wave}
\usage{
compute_fundamental_cycle(
  x,
  dimension,
  uncertainty,
  integer_harmonics_tolerance
)
}
\arguments{
\item{x}{Spectrum representing a complex waveform}

\item{dimension}{Space or time, used to label the output}

\item{uncertainty}{Precision for creating rational approximations}

\item{integer_harmonics_tolerance}{Allowable toleance for approximating least common multiples (LCM).}
}
\value{
Estimated cycle length of the complex waveform.
}
\description{
Compute the cycle length of a complex wave
}
