% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{compute_fundamental_frequency}
\alias{compute_fundamental_frequency}
\title{Compute the fundamental temporal frequency of a complex waveform.}
\usage{
compute_fundamental_frequency(x, time_uncertainty, integer_harmonics_tolerance)
}
\arguments{
\item{x}{Wavelength spectrum that include stimulus, beat, and cochlear emission tones.}

\item{time_uncertainty}{Uncertainty factor applied when creating rational approximations for temporal frequency.}

\item{integer_harmonics_tolerance}{Allowable deviation for harmonics that are not perfect integers.}
}
\value{
Fundamental temporal frequency of a complex waveform.
}
\description{
Computes the fundamental temporal frequency from a frequency spectrum that
includes stimulus, beat, and cochlear emission tones.
}
