% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{compute_fundamental_wavenumber}
\alias{compute_fundamental_wavenumber}
\title{Compute the fundamental wavenumber of the complex waveform.}
\usage{
compute_fundamental_wavenumber(
  x,
  space_uncertainty,
  integer_harmonics_tolerance
)
}
\arguments{
\item{x}{Wavelength spectrum that include stimulus, beat, and cochlear emission tones.}

\item{space_uncertainty}{Uncertainty factor applied when creating rational approximations for spatial wavelength.}

\item{integer_harmonics_tolerance}{Allowable deviation for harmonics that are not perfect integers.}
}
\value{
Fundamental wavenumber of a complex waveform.
}
\description{
Computes the fundamental wavenumber from a wavelength spectrum that
includes stimulus, beat, and cochlear emission tones.
}
