% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_pseudo_octave}
\alias{compute_pseudo_octave}
\title{compute_pseudo_octave}
\usage{
compute_pseudo_octave(fn, f0, n)
}
\arguments{
\item{fn}{freq to eval}

\item{f0}{fundamental freq}

\item{n}{harmonic number}
}
\value{
Calculated pseudo octave
}
\description{
Find the highest fundamental freq
}
