% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{default_beat_pass_filter}
\alias{default_beat_pass_filter}
\title{Default beat pass filter}
\usage{
default_beat_pass_filter()
}
\description{
Default beat pass filter for determining which beat tones pass into the model
}
\details{
'
}
