% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{default_cochlear_amplifier_num_harmonics}
\alias{default_cochlear_amplifier_num_harmonics}
\title{Default cochlear amplification number of harmonics}
\usage{
default_cochlear_amplifier_num_harmonics()
}
\description{
Default number of harmonics emitting by the cochlea
}
\details{
'
}
