% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{default_harmonics_deviation}
\alias{default_harmonics_deviation}
\title{Default Approximate Least Common Multiple Deviation}
\usage{
default_harmonics_deviation()
}
\description{
Default harmonics_deviation for approximating the Least Common Multiple (LCM)
}
\details{
'
}
