% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{default_integer_harmonics_tolerance}
\alias{default_integer_harmonics_tolerance}
\title{Default integer harmonic tolerance}
\usage{
default_integer_harmonics_tolerance()
}
\description{
Tolerance for the case when harmonics are not perfect integers
useful for approximating the Least Common Multiple (LCM) and
to determine the pseudo octave in the case of stretched timbre
}
\details{
'
}
