% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{default_integration}
\alias{default_integration}
\title{Default Integration Time}
\usage{
default_integration()
}
\description{
Default standard_deviation for converting floating point numbers to rational fractions
}
