% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{default_minimum_amplitude}
\alias{default_minimum_amplitude}
\title{Default Minimum Amplitude}
\usage{
default_minimum_amplitude()
}
\description{
Default minimum amplitude for deciding which tones are evaluated
}
\details{
'
}
