% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{default_octave_deviation}
\alias{default_octave_deviation}
\title{Default Approximate Least Common Multiple Deviation}
\usage{
default_octave_deviation()

default_octave_deviation()
}
\description{
Default octave_deviation for approximating the Least Common Multiple (LCM)

Default minimum amplitude for deciding which tones are evaluated
}
\details{
'

'
}
