% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{default_precision}
\alias{default_precision}
\title{Default Precision}
\usage{
default_precision()
}
\description{
Default precision for converting floating point numbers to rational fractions
}
\details{
'
}
