% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{default_standard_deviation}
\alias{default_standard_deviation}
\title{Default Standard Deviation}
\usage{
default_standard_deviation()
}
\description{
Default standard_deviation for converting floating point numbers to rational fractions
}
