% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{default_tolerance}
\alias{default_tolerance}
\title{Default Tolerance}
\usage{
default_tolerance()
}
\value{
Tolerance value
}
\description{
Default tolerance discovered by comparison with large-scale behavioral results
}
