% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{default_variance}
\alias{default_variance}
\title{Default Rational Fraction Precision}
\usage{
default_variance()
}
\description{
Default variance for converting floating point numbers to rational fractions
}
\details{
See Heisenberg's uncertainty paper, translated into English by John Archibald
Wheeler and Hubert Zurek, in Quantum Theory and Measurement, Wheeler and Zurek,
eds. (Princeton: Princeton Univ. Press, 1983), pp. 62-84.

See Gabor, "Theory of Communication" 1946.

See "Uncertainty Principle Inequalities and Spectrum Estimation"
in "Recent Advances in Fourier Analysis and Its Applications"
Benedetto (1989)
5. Spectrum Estimation and the Viener-Vintner Theorem

"Aspects of Gabor analysis on locally compact abelian groups"
Karlheinz Grochenig (1998)
6.3 Uncertainty Principles and Lieb's inequalities
Equation (6.3.1)

"There are various formulations of this statement \link{CP84, DS89, Fef83, Pri83},
some of which generalize to LCA groups \link{PS88, Smi}."

CP84  M. G. Cowling and J. F. Price. Bandwidth versus time concentration:
the Heisenberg-Pauli-Weyl inequality. SIAM J. Math., 15:151 - 165, 1984.

DS89 D. L. Donoho and P. B. Stark. Uncertainty principles and signal recovery.
SIAM J. Appl. Math., 49:906 - 931, 1989.

Fef83 C.L. Fefferman. The uncertainty principle. Bull. Amer. Math. Soc.
(N.S.), 9(2), 1983.

Pri83 J. F. Price. Inequalities and local uncertainty principles.
J. Math. Phys., 24:1711 - 1714, 1983.

PS88 J.F. Price and A. Sitaram. Local uncertainty inequalities for locally
compact groups. Trans. Amer. Math. Soc., 308(1):105- 114, 1988
}
