% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{empty_wavelength_spectrum}
\alias{empty_wavelength_spectrum}
\title{Helper function to create an empty spectrum}
\usage{
empty_wavelength_spectrum()
}
\description{
Helper function to create an empty spectrum
}
