% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{format_output}
\alias{format_output}
\title{Format Output}
\usage{
format_output(x, metadata, verbose)
}
\arguments{
\item{x}{The output tibble with all the data}

\item{metadata}{Metadata that the user can include that will roundtrip from input to output}

\item{verbose}{If true then all the data will be included.}

\item{...}{parameters for hrep::sparse_fr_spectrum}
}
\description{
Allows the user to select which columns from the tibble are returned
}
