% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{gcd}
\alias{gcd}
\title{Greatest Common Divisor of Rational Numbers}
\usage{
gcd(x, precision)
}
\arguments{
\item{x}{Vector of rational numbers}

\item{precision}{Precision value for creating rational fractions}
}
\value{
The greatest common divisor of the rational numbers
}
\description{
See equation 15 in "Non-Integer Arrays for Array Signal Processing"
by Kulkarni and Vaidyanathan (2022)
}
