% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{generate_cochlea_emissions}
\alias{generate_cochlea_emissions}
\title{Generate tones emitted from the cochlea in response to frequency stimulation}
\usage{
generate_cochlea_emissions(x, sfoae_num_harmonics)
}
\arguments{
\item{x}{Wavelength and frequency spectra representing the stimulus and beat tones.}

\item{sfoae_num_harmonics}{Number of harmonics emitted from the cochlea via SFOAE.}
}
\value{
Combined frequency and wavelength spectra for the stimulus, beat tones, and tones generated by the SFOAE.
}
\description{
Generate tones emitted from the cochlea in response to frequency stimulation
}
