% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{generate_combination_tones}
\alias{generate_combination_tones}
\title{Generate combination tones}
\usage{
generate_combination_tones(x, combination_coefficients)
}
\arguments{
\item{x}{Wavelength and frequency spectra representing the stimulus tones.}

\item{combination_tones}{Number of harmonics emitted from the cochlea via SFOAE.}
}
\value{
Combined frequency and wavelength spectra for the combination tones
}
\description{
Generate combination tones
}
