% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{generate_stimulus}
\alias{generate_stimulus}
\title{Generate the stimulus tones as frequency and wavelength spectra}
\usage{
generate_stimulus(x)
}
\arguments{
\item{x}{A sparse frequency spectrum from \code{hrep}, representing the
frequency content of the stimulus tones.}
}
\value{
The frequency and wavelength spectra for the stimulus tones.
}
\description{
Generate the stimulus tones as frequency and wavelength spectra
}
