% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{integration_time}
\alias{integration_time}
\title{Integration Time}
\usage{
integration_time(dimension, uncertainty)
}
\arguments{
\item{dimension}{Space or time dimension}

\item{uncertainty}{Uncertainty product}
}
\description{
Given an uncertainty returns an integration time.
}
