% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\docType{data}
\name{MIN_AMPLITUDE}
\alias{MIN_AMPLITUDE}
\title{Major-Minor Consonance-Dissonance}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MIN_AMPLITUDE
}
\arguments{
\item{x}{Chord to analyse specified in MIDI, coerced to
hrep::sparse_fr_spectrum}

\item{metadata}{User-provided list of metadata that roundtrips with each call.
helpful for analysis and plots}

\item{verbose}{Determines the amount of data to return from chord evaluation
TRUE is all and FALSE is just major-minor, consonance-dissonance, tolerance and metadata.}

\item{min_amplitude}{An optional minimum amplitude for deciding which signals to include}

\item{spatial_tolerance}{An optional tolerance value for creating rational fractions for spatial signals}

\item{temporal_tolerance}{An optional tolerance value for creating rational fractions for temporal signals}

\item{...}{parameters for hrep::sparse_fr_spectrum}
}
\value{
Major-Minor and Consonance-Dissonance plus additional information
if verbose=TRUE
}
\description{
Provides an auditory model of music perception
}
\keyword{datasets}
