% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{minimum_amplitude}
\alias{minimum_amplitude}
\title{Default Minimum Amplitude}
\usage{
minimum_amplitude()
}
\description{
Default minimum amplitude for deciding which tones are evaluated
}
\details{
'
}
