% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{most_common_pseudo_octave}
\alias{most_common_pseudo_octave}
\title{most_common_pseudo_octave}
\usage{
most_common_pseudo_octave(pseudo_octaves)
}
\arguments{
\item{x}{Vector of floating point numbers to approximate}

\item{precision}{Precision for approximations}
}
\value{
A data frame of rational numbers and metadata
}
\description{
Approximates floating-point numbers to arbitrary precision.
}
