% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{octave_deviation}
\alias{octave_deviation}
\title{Default Approximate Least Common Multiple Deviation}
\usage{
octave_deviation()
}
\description{
Default octave_deviation for approximating the Least Common Multiple (LCM)
}
\details{
'
}
