% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pseudo_octaves}
\alias{pseudo_octaves}
\title{pseudo_octaves}
\usage{
pseudo_octaves(x, deviation)
}
\arguments{
\item{x}{Chord frequencies}

\item{deviation}{Deviation for estimating least common multiples}
}
\value{
A double of the best guess of the pseudo octave
}
\description{
Determine pseudo octave of all frequencies relative to lowest frequency
}
