% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rational_fraction}
\alias{rational_fraction}
\title{rational_fraction}
\usage{
rational_fraction(x, tolerance)
}
\arguments{
\item{x}{Number to convert to fraction}

\item{tolerance}{Tolerance for converting}
}
\value{
A ratio of num / den
}
\description{
Stern-Brocot: find the highest fundamental freq
}
