% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{rational_fraction_variance}
\alias{rational_fraction_variance}
\title{Default Rational Fraction Precision}
\usage{
rational_fraction_variance()
}
\description{
Default variance for converting floating point numbers to rational fractions
}
\details{
'
}
