% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rational_fractions}
\alias{rational_fractions}
\title{rational_fractions}
\usage{
rational_fractions(x, precision)
}
\arguments{
\item{x}{Vector of floating point numbers to approximate}

\item{precision}{Precision for approximations}
}
\value{
A data frame of rational numbers and metadata
}
\description{
Approximates floating-point numbers to arbitrary precision.
}
