% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ratios}
\alias{ratios}
\title{ratios}
\usage{
ratios(x, tolerance)
}
\arguments{
\item{x}{Vector of numbers to find ratios}

\item{tolerance}{Tolerance for ratio search}
}
\value{
A data frame of ratios and metadata
}
\description{
Finds ratios within a tolerance
}
