% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stern_brocot}
\alias{stern_brocot}
\title{stern_brocot}
\usage{
stern_brocot(lower_bound, x, upper_bound)
}
\arguments{
\item{lower_bound}{Binary search stops once the approximation is above the lower bound and..,}

\item{x}{Number to convert to rational fraction}

\item{upper_bound}{... below the upper bound.}
}
\value{
A ratio of num / den
}
\description{
Approximates a floating-point number to arbitrary variance.
}
