% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{stimulus}
\alias{stimulus}
\title{Generate the stimulus wave patterns in terms of frequency spectrum and wavelength spectrum}
\usage{
stimulus(x, include_beats)
}
\arguments{
\item{x}{A sparse frequency spectrum}

\item{include_beats}{Whether to include beats in the wavelength spectrum}
}
\value{
Frequency spectrum, wavelength spectrum and wavelength beats spectrum
}
\description{
Generate the stimulus wave patterns in terms of frequency spectrum and wavelength spectrum
}
