% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{uncertainty}
\alias{uncertainty}
\title{Uncertainty}
\usage{
uncertainty(dimension, integration_time)
}
\arguments{
\item{dimension}{Space or time dimension}

\item{integration_time}{Integration time to sample the space or time signal}
}
\description{
The uncertainty product of the given dimension is used to convert
floating-point numbers to rational fractions.
}
