% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{uncertainty_limit}
\alias{uncertainty_limit}
\title{Uncertainty limit}
\usage{
uncertainty_limit()
}
\description{
Uncertainty limit for the uncertainty products of
time, frequency or space, wavelength
}
