% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{alcd}
\alias{alcd}
\title{Approximate Least Common Denominator}
\usage{
alcd(x, precision)
}
\arguments{
\item{x}{Vector of rational numbers}

\item{precision}{Precision value for creating rational fractions}
}
\value{
The greatest common divisor of the rational numbers
}
\description{
"A Practical Fundamental Frequency Extraction Algorithm
for Motion Parameters Estimation of Moving Targets"
Huang et al. (2014)
III. EXTRACTION OF A FUNDAMENTAL FREQUENCY
A. Extraction of the Initial Pitch Using the AGCD Method
}
\details{
"An Efficient Technique for Modeling and Synthesis of Automotive Engine Sounds"
Amman and Das (2001)
}
