% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{alcd}
\alias{alcd}
\title{Approximate Least Common Denominator}
\usage{
alcd(x, variance, octave_deviation, label)
}
\arguments{
\item{x}{Vector of rational numbers}

\item{variance}{Precision for creating rational fractions}

\item{octave_deviation}{Deviation for approximating least common multiples}

\item{label}{A custom label for the output usually 'period' or 'frequency'}
}
\value{
The approximate least common denominator of the rational numbers
}
\description{
See the articles below for previous work on approximate GCD / LCM
}
\details{
"Pitch extraction from corrupted harmonics of the power spectrum"
II. APPROXIMATE GCD ALGORITHM
APPENDIX B: APPROXIMATE GCD ALGORITHM
Sreenivas and Rao (1978)

"An Efficient Technique for Modeling and Synthesis of Automotive Engine Sounds"
IV. EXTRACTION OF DETERMINISTIC COMPONENTS USING AN SDFT
Amman and Das (2001)

"A Practical Fundamental Frequency Extraction Algorithm
for Motion Parameters Estimation of Moving Targets"
Huang et al. (2014)
III. EXTRACTION OF A FUNDAMENTAL FREQUENCY
A. Extraction of the Initial Pitch Using the AGCD Method
}
