% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{approximate_rational_fractions}
\alias{approximate_rational_fractions}
\title{approximate_rational_fractions}
\usage{
approximate_rational_fractions(x, variance, deviation)
}
\arguments{
\item{x}{Vector of floating point numbers to approximate}

\item{variance}{Precision for finding rational fractions}

\item{deviation}{Deviation for estimating least common multiples}
}
\value{
A data frame of rational numbers and metadata
}
\description{
Approximates floating-point numbers to arbitrary variance.
}
