% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{approximate_rational_fractions}
\alias{approximate_rational_fractions}
\title{approximate_rational_fractions}
\usage{
approximate_rational_fractions(x, reference, sd, approximate_lcm_sd)
}
\arguments{
\item{x}{Vector of floating point numbers to approximate}

\item{reference}{This will be the denominator of the ratios}

\item{sd}{Standard deviation for each evaluated parameter}

\item{approximate_lcm_sd}{Standard deviation for estimating least common multiples}
}
\value{
A data frame of rational numbers and metadata
}
\description{
Approximates floating-point numbers to arbitrary variance.
}
