% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{compute_consonance}
\alias{compute_consonance}
\title{About the speed of sound}
\usage{
compute_consonance(
  x,
  amplitude,
  temporal_variance,
  spatial_variance,
  deviation
)
}
\description{
We are dealing with ratios. Mathematically the speed of sound constant will
disappear. We could estimate the speed of sound in:
\itemize{
\item room temperature air at sea level (343 m/s)
\item the fluid of the inner ear (1,522 m/s, ocean water at room temperature)
\item the basilar membrane (1,640 m/s, human cartilage)
No matter which one we pick we would get the same mathematical results.
}
}
\details{
However, this is a computational model so we cannot avoid thinking about:
\itemize{
\item normalizing the values of the spatial and temporal signals
\item floating-point arithmetic
}

If we choose any of the physical values listed above, our wavelength numbers
computed from c_sound / f will be orders of magnitude smaller than our
frequency numbers. For example, the 10th harmonic of Eb5 has a frequency of
6,222.540 Hz. In room temperature air, the wavelength would 0.551 meters.

When we find rational fractions based on those two sets of values in the
Stern-Brocot tree, we would have to be thoughtful about the variance value.

To workaround being thoughtful about the Stern-Brocot tree, we pick a speed of
sound that is computationally friendly:

c_sound = max(f) / max(1/f)

For Eb5 with 10 harmonics, our speed of sound would be

max(Eb5_f) / max(1/Eb5_f) -> 1,627,975 m / s

That is a fast material. The advantage, though, is that our wavelength and
frequency numbers will be in the same range.

Eb5_f = 622.254 1244.508 1866.762 2489.016 3111.270 3733.524 4355.778 4978.032 5600.286 6222.540
Eb5_l = 6222.5396 3111.2698 2074.1799 1555.6349 1244.5079 1037.0899  888.9342  777.8175  691.3933  622.2540

An emergent benefit of this approach is that from a quick scan of the numbers
we can see that the two vectors are not identical.

Even with a very high variance conversion between wavelengths and frequencies,
the two periodicity estimates for the same set of waves will be different.
This is our first glimpse at the Heisenberg-Gabor uncertainty principle.

see:
https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6662181/#:~:text=The\%20speed\%20of\%20sound\%20in,\%2Fs)\%20for\%20image\%20reconstruction.
https://itis.swiss/virtual-population/tissue-properties/database/acoustic-properties/speed-of-sound/
https://www.engineeringtoolbox.com/sound-speed-water-d_598.html
https://www.engineeringtoolbox.com/air-speed-sound-d_603.html
https://en.wikipedia.org/wiki/Floating-point_arithmetic
}
