% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{compute_consonance}
\alias{compute_consonance}
\title{About the speed of sound}
\usage{
compute_consonance(x, minimum_amplitude, octave_deviation)
}
\description{
We are dealing with ratios. The speed of sound constant will
disappear. We could estimate the speed of sound as:
\itemize{
\item room temperature air at sea level (343 m/s)
\item the fluid of the inner ear (1,522 m/s, ocean water at room temperature)
\item the basilar membrane (1,640 m/s, human cartilage)
No matter which one we pick we would get the same consonance results.
}
}
\details{
Consider a speed of sound that is normalized between the two signals:

c_sound = max(f) / max(1/f)

For a chord = 100Hz pitch with 10 harmonics, our speed of sound would be

max(chord) / max(1/chord) -> 1e+05

That is a fast material. The advantage, though, is that our wavelength and
frequency numbers will be in the same range.

frequencies: 100  200    300    400    500    600    700    800    900    1000
periods: 100  111.11 125.00 142.85 166.66 200.00 250.00 333.33 500.00 1000

We can see that the two vectors are not identical.

The two periodicity estimates for the same set of waves will be different.
This is our first glimpse at the Heisenberg-Gabor uncertainty principle.

see:
https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6662181/#:~:text=The\%20speed\%20of\%20sound\%20in,\%2Fs)\%20for\%20image\%20reconstruction.
https://itis.swiss/virtual-population/tissue-properties/database/acoustic-properties/speed-of-sound/
https://www.engineeringtoolbox.com/sound-speed-water-d_598.html
https://www.engineeringtoolbox.com/air-speed-sound-d_603.html
https://en.wikipedia.org/wiki/Floating-point_arithmetic
}
