% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{cycles}
\alias{cycles}
\title{Approximate Number of Cycles of a Complex Wave}
\usage{
cycles(x, uncertainty, integer_harmonics_tolerance, label)
}
\arguments{
\item{x}{Vector of rational numbers}

\item{uncertainty}{Precision for creating rational fractions}

\item{integer_harmonics_tolerance}{Deviation for approximating least common multiples}

\item{label}{A custom label for the output usually 'space' or 'time'}
}
\value{
Estimate the number of cycles using approximate least common denominator of the rational numbers
}
\description{
Approximate Number of Cycles of a Complex Wave
}
