% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{cycles}
\alias{cycles}
\title{Approximate Number of Cycles of a Complex Wave}
\usage{
cycles(x, dimension, integration_time, harmonics_deviation)
}
\arguments{
\item{x}{Vector of rational numbers}

\item{dimension}{Space or time dimension}

\item{integration_time}{Integration time to sample the space or time signal}

\item{harmonics_deviation}{Deviation for approximating least common multiples}
}
\value{
Estimate the number of cycles using approximate least common denominator of the rational numbers
}
\description{
Approximate Number of Cycles of a Complex Wave
}
