% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{default_integration_time}
\alias{default_integration_time}
\title{Default Integration Time}
\usage{
default_integration_time()
}
\description{
Default integration time for sampling the space and time signals
}
