% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{mami.codi}
\alias{mami.codi}
\title{Major-Minor Consonance-Dissonance}
\usage{
mami.codi(
  x,
  amplitude_lower_bound = AMPLITUDE_LOWER_BOUND_DEFAULT,
  temporal_frequency_sd = NA,
  spatial_frequency_sd = NA,
  approximate_lcm_sd = APPROXIMATE_LCM_SD_DEFAULT,
  metadata = NA,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{Chord to analyse specified in MIDI, coerced to
hrep::sparse_fr_spectrum}

\item{amplitude_lower_bound}{An optional minimum amplitude_lower_bound for deciding which partials to include.}

\item{temporal_frequency_sd}{An optional temporal variance value for finding rational fractions.}

\item{spatial_frequency_sd}{An optional spatial variance value for finding rational fractions.}

\item{approximate_lcm_sd}{An optional approximate_lcm_sd value for approximating least common multiples.}

\item{metadata}{User-provided list of metadata that round trips with each call.
helpful for analysis and plots}

\item{verbose}{Determines the amount of data to return from chord evaluation
TRUE is all and FALSE is just major-minor, consonance-dissonance and metadata.}

\item{...}{parameters for hrep::sparse_fr_spectrum}
}
\value{
Major-Minor, Consonance-Dissonance and metadata plus additional
information if verbose=TRUE
}
\description{
A spatiotemporal periodicity model of consonance perception.
}
