% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{mami.codi}
\alias{mami.codi}
\title{Major-Minor Consonance-Dissonance}
\usage{
mami.codi(
  x,
  minimum_amplitude = MINIMUM_AMPLITUDE,
  time_standard_deviation = STANDARD_DEVIATION,
  space_standard_deviation = STANDARD_DEVIATION,
  harmonics_deviation = HARMONICS_DEVIATION,
  metadata = NA,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{Chord to analyse specified in MIDI, coerced to
hrep::sparse_fr_spectrum}

\item{time_standard_deviation}{An optional time standard_deviation value for finding rational fractions.}

\item{space_standard_deviation}{An optional space standard_deviation value for finding rational fractions.}

\item{harmonics_deviation}{An optional deviation value that allows for harmonics to be not perfect integers.}

\item{metadata}{User-provided list of metadata that round trips with each call.
helpful for analysis and plots}

\item{verbose}{Determines the amount of data to return from chord evaluation
TRUE is all and FALSE is just major-minor, consonance-dissonance and metadata.}

\item{...}{parameters for hrep::sparse_fr_spectrum}

\item{amplitude}{An optional minimum amplitude for deciding which partials to include.}
}
\value{
Major-Minor (majorness), Consonance-Dissonance (dissonance) and metadata plus additional
information if verbose=TRUE
}
\description{
A model of consonance perception based on space and time cycles.
}
