% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{mami.codi}
\alias{mami.codi}
\title{Major-Minor Consonance-Dissonance}
\usage{
mami.codi(
  x,
  amplitude = MINIMUM_AMPLITUDE,
  temporal_variance = NA,
  spatial_variance = NA,
  deviation = OCTAVE_DEVIATION,
  metadata = NA,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{Chord to analyse specified in MIDI, coerced to
hrep::sparse_fr_spectrum}

\item{amplitude}{An optional minimum amplitude for deciding which partials to include.}

\item{temporal_variance}{An optional temporal variance value for finding rational fractions.}

\item{spatial_variance}{An optional spatial variance value for finding rational fractions.}

\item{deviation}{An optional deviation value for approximating least common multiples.}

\item{metadata}{User-provided list of metadata that round trips with each call.
helpful for analysis and plots}

\item{verbose}{Determines the amount of data to return from chord evaluation
TRUE is all and FALSE is just major-minor, consonance-dissonance and metadata.}

\item{...}{parameters for hrep::sparse_fr_spectrum}
}
\value{
Major-Minor, Consonance-Dissonance and metadata plus additional
information if verbose=TRUE
}
\description{
A spatiotemporal periodicity model of consonance perception.
}
