% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{mami.codi}
\alias{mami.codi}
\title{Major-Minor Consonance-Dissonance}
\usage{
mami.codi(
  x,
  beat_pass_filter = BEAT_PASS_FILTER$ALL,
  cochlear_amplifier_num_harmonics = COCHLEAR_AMPLIFIER_NUM_HARMONICS,
  space_uncertainty = UNCERTAINTY_LIMIT,
  time_uncertainty = UNCERTAINTY_LIMIT,
  integer_harmonics_tolerance = INTEGER_HARMONICS_TOLERANCE,
  metadata = NA,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{Chord to analyze, which is parsed using
\code{hrep::sparse_fr_spectrum}. For more details, see
\href{https://github.com/pmcharrison/hrep/blob/master/R/sparse-fr-spectrum.R}{hrep::sparse_fr_spectrum documentation}.}

\item{beat_pass_filter}{Specifies which bands of beats to include: low, high, all, or none.
\describe{
@eval paste0(
"\item{\\code{", BEAT_PASS_FILTER$LOW, "}}{Passes beat tones lower than the lowest stimulus tone.}\n",
"\item{\\code{", BEAT_PASS_FILTER$HIGH, "}}{Passes beat tones higher than the lowest stimulus tone.}\n",
"\item{\\code{", BEAT_PASS_FILTER$ALL, "}}{Passes all beat tones.}\n",
"\item{\\code{", BEAT_PASS_FILTER$NONE, "}}{Passes none of the beat tones.}\n"
)
}}

\item{cochlear_amplifier_num_harmonics}{Number of harmonics to include for the stimulus-frequency otoacoustic emission.}

\item{space_uncertainty}{Optional space uncertainty value for finding rational fractions.}

\item{time_uncertainty}{Optional time uncertainty value for finding rational fractions.}

\item{integer_harmonics_tolerance}{Optional tolerance for harmonics to deviate from perfect integers.}

\item{metadata}{User-provided list of metadata that accompanies each call. Useful for analysis and plots.}

\item{verbose}{Determines the amount of data to return from chord evaluation. Options are:
\describe{
\item{\code{TRUE}}{Returns all available evaluation data.}
\item{\code{FALSE}}{Returns only essential data: major/minor classification, consonance/dissonance, and metadata.}
}}

\item{...}{Additional parameters for \code{hrep::sparse_fr_spectrum}.}
}
\value{
Major-Minor (majorness), Consonance-Dissonance (dissonance), and metadata, with additional
information if \code{verbose = TRUE}.
}
\description{
A model of consonance perception based on the fundamental cycles of complex space and time signals.
}
