% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{mami.codi}
\alias{mami.codi}
\title{Major-Minor Consonance-Dissonance}
\usage{
mami.codi(
  x,
  sfoae_num_harmonics = 5,
  include_beats = T,
  minimum_amplitude = MINIMUM_AMPLITUDE,
  integration_time = INTEGRATION_TIME,
  harmonics_deviation = HARMONICS_DEVIATION,
  metadata = NA,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{Chord to analyse specified in MIDI, coerced to
hrep::sparse_fr_spectrum}

\item{sfoae_num_harmonics}{Stimulus Frequency Otoacoustic Emissions}

\item{include_beats}{Beats}

\item{minimum_amplitude}{An optional minimum amplitude for deciding which partials to include.}

\item{integration_time}{Integration time to sample the space and time signals}

\item{harmonics_deviation}{An optional deviation value that allows for harmonics to be not perfect integers.}

\item{metadata}{User-provided list of metadata that round trips with each call.
helpful for analysis and plots}

\item{verbose}{Determines the amount of data to return from chord evaluation
TRUE is all and FALSE is just major-minor, consonance-dissonance and metadata.}

\item{...}{parameters for hrep::sparse_fr_spectrum}
}
\value{
Major-Minor (majorness), Consonance-Dissonance (dissonance) and metadata plus additional
information if verbose=TRUE
}
\description{
A model of consonance perception based on space and time cycles.
}
