% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{mami.codi}
\alias{mami.codi}
\title{Major-Minor Consonance-Dissonance}
\usage{
mami.codi(
  x,
  minimum_amplitude = MINIMUM_AMPLITUDE,
  temporal_standard_deviation = STANDARD_DEVIATION,
  spatial_standard_deviation = STANDARD_DEVIATION,
  harmonics_deviation = HARMONICS_DEVIATION,
  metadata = NA,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{Chord to analyse specified in MIDI, coerced to
hrep::sparse_fr_spectrum}

\item{temporal_standard_deviation}{An optional temporal standard_deviation value for finding rational fractions.}

\item{spatial_standard_deviation}{An optional spatial standard_deviation value for finding rational fractions.}

\item{harmonics_deviation}{An optional deviation value for approximating least common multiples.}

\item{metadata}{User-provided list of metadata that round trips with each call.
helpful for analysis and plots}

\item{verbose}{Determines the amount of data to return from chord evaluation
TRUE is all and FALSE is just major-minor, consonance-dissonance and metadata.}

\item{...}{parameters for hrep::sparse_fr_spectrum}

\item{amplitude}{An optional minimum amplitude for deciding which partials to include.}
}
\value{
Major-Minor, Consonance-Dissonance and metadata plus additional
information if verbose=TRUE
}
\description{
A spatiotemporal periodicity model of consonance perception.
}
