% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{mami.codi}
\alias{mami.codi}
\title{Major-Minor Consonance-Dissonance}
\usage{
mami.codi(
  x,
  minimum_amplitude = MINIMUM_AMPLITUDE,
  precision = RATIONAL_FRACTION_PRECISION,
  metadata = NA,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{Chord to analyse specified in MIDI, coerced to
hrep::sparse_fr_spectrum}

\item{minimum_amplitude}{An optional minimum amplitude for deciding which signals to include}

\item{precision}{An optional precision value for creating rational fractions}

\item{metadata}{User-provided list of metadata that roundtrips with each call.
helpful for analysis and plots}

\item{verbose}{Determines the amount of data to return from chord evaluation
TRUE is all and FALSE is just major-minor, consonance-dissonance, precision and metadata.}

\item{...}{parameters for hrep::sparse_fr_spectrum}
}
\value{
Major-Minor and Consonance-Dissonance plus additional information
if verbose=TRUE
}
\description{
Provides an auditory model of music perception
}
