% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{mami.codi}
\alias{mami.codi}
\title{Major-Minor Consonance-Dissonance}
\usage{
mami.codi(
  x,
  min_amplitude = MIN_AMPLITUDE,
  spatial_tolerance = TOLERANCE,
  temporal_tolerance = TOLERANCE,
  metadata = NA,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{Chord to analyse specified in MIDI, coerced to
hrep::sparse_fr_spectrum}

\item{min_amplitude}{An optional minimum amplitude for deciding which signals to include}

\item{spatial_tolerance}{An optional tolerance value for creating rational fractions for spatial signals}

\item{temporal_tolerance}{An optional tolerance value for creating rational fractions for temporal signals}

\item{metadata}{User-provided list of metadata that roundtrips with each call.
helpful for analysis and plots}

\item{verbose}{Determines the amount of data to return from chord evaluation
TRUE is all and FALSE is just major-minor, consonance-dissonance, tolerance and metadata.}

\item{...}{parameters for hrep::sparse_fr_spectrum}
}
\value{
Major-Minor and Consonance-Dissonance plus additional information
if verbose=TRUE
}
\description{
Provides an auditory model of music perception
}
