% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pseudo_octaves}
\alias{pseudo_octaves}
\title{pseudo_octaves}
\usage{
pseudo_octaves(x)
}
\arguments{
\item{x}{Chord frequencies}
}
\value{
A double of the best guess of the pseudo octave
}
\description{
Determine pseudo octave of all frequencies relative to lowest frequency
}
