% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{rational_fraction_precision}
\alias{rational_fraction_precision}
\title{Default Rational Fraction Precision}
\usage{
rational_fraction_precision()
}
\description{
Default precision for converting floating point numbers to rational fractions
}
\details{
'
}
