% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{space_time_cycles}
\alias{space_time_cycles}
\title{Compute the cycle lengths from space and time signals}
\usage{
space_time_cycles(
  x,
  minimum_amplitude,
  time_uncertainty,
  space_uncertainty,
  integer_harmonics_tolerance
)
}
\arguments{
\item{x}{Stimulus encoded as frequency and wavelength spectra}

\item{minimum_amplitude}{Ignore all frequencies below this threshold}

\item{time_uncertainty}{Uncertainty for creating rational fractions}

\item{space_uncertainty}{Uncertainty for creating rational fractions}

\item{integer_harmonics_tolerance}{Error allowance for harmonics not being perfect integers}
}
\value{
Estimate the cycle length of time and space signals
}
\description{
Compute the cycle lengths from space and time signals
}
