% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stern_brocot}
\alias{stern_brocot}
\title{stern_brocot}
\usage{
stern_brocot(x, standard_deviation)
}
\arguments{
\item{x}{Number to convert to rational fraction}

\item{standard_deviation}{Binary search stops once the desired standard_deviation is reached}
}
\value{
A ratio of num / den
}
\description{
Approximates a floating-point number to arbitrary standard_deviation.
}
